// src/app/shared/models/module-control.ts
export interface ModuleControl {
   moduleControlId: number
   ctrlType: string
   moduleCode: string
   appCode: string
   ctrlName: string
   ctrlDescription: string
   ctrlValue: string
   sequence: number
   createdById: number
   createdBy: string
   createdAt: Date
   modifiedById: number
   modifiedBy: string
   modifiedAt: Date
   deactivated: boolean
 }
 